ref: class#Product change price type to BigDecimal

+++ product-service/src/main/java/com/arturfrimu/product/dto/command/CreateProductCommand.java
@@ -3,14 +3,16 @@ package com.arturfrimu.product.dto.command;
 import com.arturfrimu.product.dto.request.CreateProductRequest;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public final class CreateProductCommand {
     private final String name;
     private final String description;
-    private final Double price;
+    private final BigDecimal price;
     private final Long categoryId;
 
-    public CreateProductCommand(String name, String description, Double price, Long categoryId) {
+    public CreateProductCommand(String name, String description, BigDecimal price, Long categoryId) {
         this.name = name;
         this.description = description;
         this.price = price;
